+++ modules/core/services/src/main/scala/org/corespring/services/item/ItemService.scala
@@ -38,7 +38,7 @@ trait ItemService extends BaseContentService[Item, VersionedId[ObjectId]] {
 
   def saveNewUnpublishedVersion(id: VersionedId[ObjectId]): Option[VersionedId[ObjectId]]
 
-  def count(query: DBObject, fields: Option[String] = None): Int
+  def count(query: DBObject, fields: Option[String] = None): Long
 
   def findFieldsById(id: VersionedId[ObjectId], fields: DBObject = MongoDBObject.empty): Option[DBObject]
